+++ core-contracts/StakedLP/build.gradle
@@ -27,6 +27,7 @@ dependencies {
 
     testImplementation 'foundation.icon:javaee-unittest:0.9.2'
     testImplementation 'com.github.sink772:javaee-tokens:0.6.1'
+    testImplementation 'org.json:json:20180130'
     // Use JUnit Jupiter for testing.
     testImplementation 'org.junit.jupiter:junit-jupiter:5.8.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
